# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Credit Service API
  version: 1.0.0
  description: API for managing bank customers

servers:
  - url: http://localhost/api/v1

paths:
  /credits/credit:
    post:
      operationId: createCredit
      summary: Create a new customer credit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCreateRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

  /credits/credit/by-customer/{customerId}:
    get:
      operationId: getCreditsByCustomerId
      summary: Get all credits by customer id
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of credits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

  /credits/credit/{id}:
    get:
      operationId: getCreditById
      summary: Get a credit by its id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'
    patch:
      operationId: updateCreditById
      summary: Update credit by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPatchRequest'
      responses:
        '200':
          description: Credit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponse'
        '404':
          description: Credit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'


  /credits/credit-card:
    post:
      operationId: createCreditCredit
      summary: Create a new customer credit card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardCreateRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

  /credits/credit-card/{id}:
    patch:
      operationId: updateCreditCard
      summary: Update credit card info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardPatchRequest'
      responses:
        '200':
          description: Credit card updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

  /credits/credit-card/by-customer/{customerId}:
    get:
      operationId: getCreditCardsByCustomerId
      summary: Get all credit cards of the customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of credit cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

  /credits/credit-card/by-card-number/{cardNumber}:
    get:
      operationId: getCreditCardByCardNumber
      summary: Get a credit card by a card number
      parameters:
        - name: cardNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Credit card found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponse'
        '404':
          description: Credit card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

components:
  schemas:
    CreditCreateRequest:
      type: object
      properties:
        customerId:
          type: string
        amount:
          type: number
        status:
          type: string
        interestRate:
          type: number
        termInMonths:
          type: integer
      required:
        - customerId
        - amount
        - interestRate
        - termInMonths

    CreditResponse:
      type: object
      properties:
        clientId:
          type: string
        amount:
          type: number
        status:
          type: string
        interestRate:
          type: number
        termInMonths:
          type: integer
        monthlyPayment:
          type: number
        amountPaid:
          type: number
        totalAmount:
          type: number

    CreditCardCreateRequest:
      type: object
      properties:
        customerId:
          type: string
        creditLimit:
          type: number
      required:
        - customerId
        - creditLimit

    CreditCardPatchRequest:
      type: object
      properties:
        usedAmount:
          type: number
        creditLimit:
          type: number

    CreditPatchRequest:
      type: object
      properties:
        amountPaid:
          type: number
      required:
        - amountPaid

    CreditCardResponse:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        creditLimit:
          type: number
        cardNumber:
          type: string
        usedAmount:
          type: number
        creditCardStatus:
          type: string

    CustomerResponse:
      type: object
      properties:
        id:
          type: string
        documentNumber:
          type: string
        ruc:
          type: string
        businessName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        type:
          type: string
        status:
          type: string
        subType:
          type: string

    ResponseBase:
      type: object
      properties:
        message:
          type: string